Consolidate backend and retention modules
