+++ steps_summary.md
@@ -137,3 +137,4 @@
 
 
 
+- Consolidated hardware backends and retention modules into single shared files.
